requirements.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
pydantic
python-dotenv
requests
pdfplumber
python-multipart
redis